refactor checkin
